Fixed several minor warnings/missing includes discovered by building with more modern toolchains. Mostly printf format specifiers, signed/unsigned comparisons and missing includes

+++ libraries/AudioFile/AudioFile.cpp
@@ -251,5 +251,5 @@ void AudioFileWriter::io(std::vector<float>& buffer)
 	size_t count = buffer.size();
 	sf_count_t ret = sf_write_float(sndfile, buffer.data(), buffer.size());
 	if(ret != sf_count_t(buffer.size()))
-		fprintf(stderr, "Error while writing to file: %lld\n", ret);
+		fprintf(stderr, "Error while writing to file: %lu\n", (long unsigned int)ret);
 }
